Initial Commit of Slides for ACCU 2021 Talk
This talk is called 'C++20 + Lua = Flexibility' and provides
an up-to-date tutorial on how to combine modern C++ with Lua.
In particular, the talk is intended to help individuals who
would like to experiment with this technique, but have no idea
where to start and are confused by the wide range of available
technologies. The talk focuses on sol3 and SWIG.