Updated PDF Export Code
This commit updates the pdf export code. This allows the user to obtain
a version of the presentation suitable for printing by going to the
URL of the talk and adding /?print-pdf. The slides are then suitable
for printing. Note that in Chrome, you need to open the print dialog
box (i.e. Ctrl+P) and select Landscape mode, with no margins and enable
background images. The file can then be saved as a pdf or printed.
Note that this has only been tested in Chrome.