Add Sol3 Code Example and Tidy-up
This commit adds a code example for Sol3. The example performs
the same function as lua-cpp.cpp, but demonstrates the more
succint and C++ friendly syntax.

This commit also tidies up the 'old' lua-example directory, which
contained obsolete content.

Note also that the sol2 tree is included as a git submodule.