Add Title Slide and Code to Hide Controls
Jim Roper (digital-medium.co.uk/ACCU) has very kindly supplied
me with a title slide for the presentation. The rationale being
that all of the talks start with a uniform look-and-feel. This
commit adds the title slide to the repository with some JS code
to hide the presentation's controls (on the title slide).