Fixes for the 'Title Slide' Code
The code to re-enable the controls for the Menu and Reveal
navigation did not trigger in a number of cases. For example,
if the presentation was started in the middle of the slides,
then the event to re-enable the controls does not fire (as this
event is only launched from the second slide).